requirements.txt
asn1crypto==0.23.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cryptography>=2.3
defusedxml==0.5.0
Django==2.0.3
django-cors-headers==2.1.0
django-encrypted-model-fields==0.5.3
djangorestframework==3.7.7
djangorestframework-expiring-authtoken
gunicorn==19.6.0
idna==2.5
jira==1.0.10
jsonfield==2.0.2
keyring==10.4.0
numpy==1.13.1
oauthlib==2.0.2
pandas==0.20.3
pbr==3.1.1
pycparser==2.18
pycrypto==2.6.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.2
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
scipy==0.19.1
six==1.10.0
urllib3==1.22
gevent==1.3.4
